Cleanup after PR feedback
