+++ src/bin-support.js
@@ -2,18 +2,13 @@
 
 const DEFAULT_JS_EXTENSIONS = 'js,ts';
 
-function getTransformPath(binRoot, transformName) {
-  const path = require('path');
-
-  return path.join(binRoot, '..', 'transforms', transformName, 'index.js');
-}
-
-async function runJsTransform(binRoot, transformName, args, extensions = DEFAULT_JS_EXTENSIONS) {
+async function runJsTransform(root, transformName, args, extensions = DEFAULT_JS_EXTENSIONS) {
   const globby = require('globby');
   const execa = require('execa');
   const chalk = require('chalk');
   const path = require('path');
   const { parseTransformArgs } = require('./options-support');
+  const { getTransformPath } = require('./transform-support');
 
   let { paths, options } = parseTransformArgs(args);
 
@@ -21,7 +16,7 @@ async function runJsTransform(binRoot, transformName, args, extensions = DEFAULT
     let foundPaths = await globby(paths, {
       expandDirectories: { extensions: extensions.split(',') },
     });
-    let transformPath = getTransformPath(binRoot, transformName);
+    let transformPath = getTransformPath(root, transformName);
 
     let jscodeshiftPkg = require('jscodeshift/package');
     let jscodeshiftPath = path.dirname(require.resolve('jscodeshift/package'));
@@ -43,15 +38,16 @@ async function runJsTransform(binRoot, transformName, args, extensions = DEFAULT
   }
 }
 
-async function runTemplateTransform(binRoot, transformName, args) {
+async function runTemplateTransform(root, transformName, args) {
   const execa = require('execa');
   const chalk = require('chalk');
   const { parseTransformArgs } = require('./options-support');
+  const { getTransformPath } = require('./transform-support');
 
   let { paths, options } = parseTransformArgs(args);
 
   try {
-    let transformPath = getTransformPath(binRoot, transformName);
+    let transformPath = getTransformPath(root, transformName);
     let binOptions = ['-t', transformPath, ...paths];
 
     return execa('ember-template-recast', binOptions, {
@@ -70,16 +66,18 @@ async function runTemplateTransform(binRoot, transformName, args) {
 }
 
 async function runTransform(binRoot, transformName, args, extensions) {
-  const { getTransformType } = require('./transform-support');
+  const { getTransformType, getTransformPath } = require('./transform-support');
+  const path = require('path');
 
-  let transformPath = getTransformPath(binRoot, transformName);
+  let root = path.join(binRoot, '..');
+  let transformPath = getTransformPath(root, transformName);
   let type = getTransformType(transformPath);
 
   switch (type) {
     case 'js':
-      return runJsTransform(binRoot, transformName, args, extensions);
+      return runJsTransform(root, transformName, args, extensions);
     case 'hbs':
-      return runTemplateTransform(binRoot, transformName, args);
+      return runTemplateTransform(root, transformName, args);
     default:
       throw new Error(`Unknown type passed to runTransform: "${type}"`);
   }
